Impute Missing Values

Impute Missing Values

impute_var(
  data,
  id_site = NULL,
  id_year = NULL,
  id_impute = NULL,
  method = "miceranger",
  n_impute = 5,
  carryforward_yrs = NULL,
  ...
)

Arguments

data

data.frame

id_site

A site-level variable name included in data.

id_year

(Default = NULL). A year variable name included in data. If not specified, assumes the dataset is cross-sectional.

id_impute

(Default = NULL) Vector with one or more variable names for which imputation is performed. Imputes all variables that contain missing values in data except id_site and id_year if not specified.

method

(Optional. Default = "miceranger") Imputation method. Choose from "carryforward", "amelia", and "miceranger". "carryforward" is available only for time-series data.

n_impute

(Optional. Used only when "miceranger" or "amelia" is selected in method. Default = 5) The number of imputed datasets to create (equivalent of argument m in amelia() and miceRanger()).

carryforward_yrs

(Optional. Used only when "carryforward" is selected in method. Default = NULL) Maximum number of years allowed to be carried forward. If not specified, it will impute values from the most recent years available.

...

Arguments passed onto Amelia::amelia() or miceRanger::miceRanger().

Value

A dataframe with new variables merged onto the base dataset.

References

Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at https://naokiegami.com/paper/sps.pdf.